feat: Update image source and button routing
This commit updates the image source in the App component to use a new image file "img1Women.jpg" instead of "img1.jpg". Additionally, it modifies the ButtonPerso component to use the "fcRuta" prop for routing instead of the "ruta" prop. The "goPersonalizar" function is also updated to use the "router.push" method for navigation. These changes improve the visual appearance and functionality of the app.