chore: Update app.json, package.json, and eas.json
This commit updates the app.json file to add the "bundleIdentifier" field for iOS and the "package" field for Android. It also modifies the package.json file to update the "start" script to use the "--dev-client" flag and changes the "main" entry point to "expo-router/entry". Additionally, it adds new dependencies for "expo-constants" and "react-dom". Finally, it introduces a new eas.json file with configuration for the EAS project ID.

```